+++ addOns/pscanrules/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - 'Information Disclosure: Suspicious Comments' if matched within script block or JS response raise Alert with Low confidence.
 - Migrate an input file from Beta to Release that were missed during previous promotions.
   - This addresses errors such as `[ZAP-PassiveScanner] ERROR org.zaproxy.zap.extension.pscanrules.InformationDisclosureInURL  - No such file: .... /xml/URL-information-disclosure-messages.txt`
+- 'Application Error' scan rule now supports custom payloads when used in conjunction with the Custom Payloads addon.
 
 ## [26] - 2020-01-17